UIDS-18 Add a wrapper to the ReactSelect's Async
This adds a wrapper to the Async component from React Select. This adds React
Select as a dependency of this library as well as brings over some styling we
had on it from the main repo. Additionally this adds a colors.js file to allow
us to reference our colors in JS for places where we need to use CSS in JS. Also
this adds the z stack variables to this repo.

Additionally this makes it so we ignore RubyMine and Intellij config files.