UIDS-18 Add aria label props and fix eslint
This adds aria-label and aria-labelledby props to the react select
components. In doing so noted that we were updating ESLint rules for
just src and not the stories so moved some relevant rules down to the
common area. Additionally added labeled components to the stories. Fixed
the lack of an import for zStack that is used on the modal imports.