UIDS-18 Move input colors to their own file.
In adding input colors separated out the color palette from other semantic color
variables. This led to moving the palette to be colors in js and scss and added
separate files for the palette (which defines the base colors we should use) as
well as other files for input colors which are joined in a common color output.