UIDS 18 react select wrappers (#19)
UIDS-18 Add a wrapper to the ReactSelect and colors

This adds a wrapper to the Async component from React Select. This adds React
Select as a dependency of this library as well as brings over some styling we
had on it from the main repo. Also add a SingleSelect component to hold logic for
that style of component. This is largely meant to help us standardize styles for these
components which use CSS in JS. 

In adding input colors separated out the color palette from other semantic color
variables. This led to moving the palette to be colors in js and scss and added
separate files for the palette (which defines the base colors we should use) as
well as other files for input colors which are joined in a common color output.

Additionally brought some ESLint rules to the top level instead of just living in src.