UIDS-506 Update Form to handle GET method
Previously passing a GET method would have worked for Rails, but it wouldn't
put the params in the URL which is the expected behavior. This makes it so we
properly pass GET directly to the form method parameter.

One other thing this does is remove the CSRF params if GET is passed. These are
ignored in GET and thus aren't necessary and would add extra params to the URL.

One potential upgrade that was left on the cutting room floor was passing any
extra props directly to the form component.