Add bugsnag config to DS and handle toast arg error (#845)
Add bugsnag configuration to DS in order to handle toast arg error.

If users make a function call to useToast with the incorrect arg signature, it will trigger a bugsnag report. Eventually we hope to remove this error handling once we are confident we have correctly updated all use cases within rails-server.