Upgrade Storybook 6.5 and Node >= 16 < 17
Loosens the rigidness of the Node engine requirement for this application.

It won’t currently work with the current Node 18, but Node 14 is out of maintenance soon and we will be needing to move off of it.

There are further deprecation issues with this currently that will need to be addressed at some point including

* Upgrading to SB 7
* Replacing `knobs` add-on
* Supporting Node 18 runtime